+++ CHANGELOG.md
@@ -6,6 +6,16 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.7.1 - Unreleased
 
+### Added
+
+* Added basic support for groups and mutually exclusive groups by specifying an id in each value.
+
+
+### Changed
+
+* Options marked as isflag can now accept any type for the key/val style of
+  argument. The flag style will still return a boolean.
+
 
 ## Version 0.7.0 - Released 2022-09-17